fixing changelog
