+++ tests/test_minimal/test_hdmf.py
@@ -1,4 +1,5 @@
 import numpy as np
+from numpy.testing import assert_array_equal
 from hdmf.testing import TestCase
 
 from neuroconv.tools.hdmf import SliceableDataChunkIterator
@@ -34,3 +35,19 @@ def test_min_axis_too_large():
     """uses ~8 MB array with each contiguous axis at around ~8 KB with 5 KB buffer_size and 1 KB chunk size."""
     iterator = SliceableDataChunkIterator(data=np.empty(shape=(1000, 1000)), chunk_mb=1e-3, buffer_gb=5e-6)
     assert iterator.buffer_shape == (22, 22)
+
+
+def test_sliceable_data_chunk_iterator():
+
+    data = np.arange(100).reshape(10, 10)
+
+    dci = SliceableDataChunkIterator(data=data, buffer_shape=(5, 5), chunk_shape=(5, 5))
+
+    data_chunk = next(dci)
+
+    assert data_chunk.selection == (slice(0, 5, None), slice(0, 5, None))
+
+    assert_array_equal(
+        data_chunk.data,
+        [[0, 1, 2, 3, 4], [10, 11, 12, 13, 14], [20, 21, 22, 23, 24], [30, 31, 32, 33, 34], [40, 41, 42, 43, 44]],
+    )
